Move logic to one service
