--- sensor-managment/src/main/java/pl/piotr/iotdbmanagement/service/MeasurementService.java
@@ -1,23 +1,0 @@
-package pl.piotr.iotdbmanagement.service;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import pl.piotr.iotdbmanagement.measurement.Measurement;
-import pl.piotr.iotdbmanagement.measurement.MeasurementRepository;
-
-@Service
-public class MeasurementService {
-    private MeasurementRepository measurementRepository;
-
-    @Autowired
-    public MeasurementService(MeasurementRepository measurementRepository) {
-        this.measurementRepository = measurementRepository;
-    }
-
-    @Transactional
-    public Measurement create(Measurement measurement) {
-        return measurementRepository.save(measurement);
-    }
-
-}
