--- sensor-managment/src/main/java/pl/piotr/iotdbmanagement/service/PlaceService.java
@@ -1,24 +1,0 @@
-package pl.piotr.iotdbmanagement.service;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import pl.piotr.iotdbmanagement.place.Place;
-import pl.piotr.iotdbmanagement.place.PlaceRepository;
-
-@Service
-public class PlaceService {
-
-    private PlaceRepository placeRepository;
-
-    @Autowired
-    public PlaceService(PlaceRepository placeRepository) {
-        this.placeRepository = placeRepository;
-    }
-
-    @Transactional
-    public Place create(Place place) {
-        return placeRepository.save(place);
-    }
-
-}
